Move algorithms to submodule (#121)
* Update imports and wording in tests.

* Update subpackage inits and util files.

* Move algorithm funcs to algorithm subpackage.

* Format codebase with isort, black, flake8.

* Update API docs.
