+++ docs/reference/source/matching.algorithms.rst
@@ -1,0 +1,54 @@
+matching.algorithms package
+===========================
+
+Submodules
+----------
+
+matching.algorithms.hospital\_resident module
+---------------------------------------------
+
+.. automodule:: matching.algorithms.hospital_resident
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+matching.algorithms.stable\_marriage module
+-------------------------------------------
+
+.. automodule:: matching.algorithms.stable_marriage
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+matching.algorithms.stable\_roommates module
+--------------------------------------------
+
+.. automodule:: matching.algorithms.stable_roommates
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+matching.algorithms.student\_allocation module
+----------------------------------------------
+
+.. automodule:: matching.algorithms.student_allocation
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+matching.algorithms.util module
+-------------------------------
+
+.. automodule:: matching.algorithms.util
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: matching.algorithms
+   :members:
+   :undoc-members:
+   :show-inheritance:
