+++ docs/reference/source/matching.rst
@@ -5,12 +5,22 @@ Subpackages
 -----------
 
 .. toctree::
+   :maxdepth: 4
 
+   matching.algorithms
    matching.games
    matching.players
 
 Submodules
 ----------
+
+matching.exceptions module
+--------------------------
+
+.. automodule:: matching.exceptions
+   :members:
+   :undoc-members:
+   :show-inheritance:
 
 matching.game module
 --------------------
