+++ src/matching/games/__init__.py
@@ -1,17 +1,13 @@
-""" Make the games and their algorithms accessible. """
+""" Top-level imports for the `matching.games` subpackage. """
 
-from .hospital_resident import HospitalResident, hospital_resident
-from .stable_marriage import StableMarriage, stable_marriage
-from .stable_roommates import StableRoommates, stable_roommates
-from .student_allocation import StudentAllocation, student_allocation
+from .hospital_resident import HospitalResident
+from .stable_marriage import StableMarriage
+from .stable_roommates import StableRoommates
+from .student_allocation import StudentAllocation
 
 __all__ = [
-    HospitalResident,
-    StableMarriage,
-    StableRoommates,
-    StudentAllocation,
-    hospital_resident,
-    stable_marriage,
-    stable_roommates,
-    student_allocation,
+    "HospitalResident",
+    "StableMarriage",
+    "StableRoommates",
+    "StudentAllocation",
 ]
