+++ src/matching/players/__init__.py
@@ -1,7 +1,7 @@
-""" Make the player classes more accessible. """
+""" Top-level imports for the `matching.players` subpackage. """
 
 from .hospital import Hospital
 from .project import Project
 from .supervisor import Supervisor
 
-__all__ = [Hospital, Project, Supervisor]
+__all__ = ["Hospital", "Project", "Supervisor"]
